Add anyhow, check credential-helper output
The program seems to advise you to run `$credential_helper login`
whenever that would help, so we do a regex search that looks to be
advising that. We do a fuzzy match in part because they do a command -v
on the credential helper and include its full path at the moment, but we
just want to search for the basename.

This also spawns threads for stdin in cases where we also read from
stderr or stdout, since otherwise that can deadlock in some cases.

For the credential-helper login step, we just pass the output straight
through.