Check remote credential (fixes #1) (#2)
Rather than checking the local credential to decide whether to login and
then unconditionally syncing the credential, we check the remote
credential to decide whether to do anything at all.

Practically speaking, this obviates needing to type your password in the
case where the remote credential is still valid.

To implement it, I wound up writing a small SSH multiplexer class. It
will start an ssh ControlMaster with a socket in a temporary directory
(unless `-r` is passed in which case it assumes you have a socket
configured already) and multiplex its SSH commands over that socket to
save a bit of time.