Add github workflow (#4)
This gets us built artifacts per OS on which someone might use this,
with repository-default retention period for pushes to main or versioned
tags. The binaries are not yet signed, but this should at least get us
started (and confirm that we build on all supported platforms.)

This also fixes a build error on Windows that was revealed by the
workflow.