Show the local error first
There are fewer moving parts on the local side, and the most likely
failure is going to be that aspect-credential-helper is not installed.
This hopefully should help the user see and address that error before
getting into things like whether the SSH configuration is working or the
VM is set up the right way.