Simplification, refactoring
has_user_socket is now called infer_create_socket, and has the same
polarity as create_socket (it returns true if the program should do
something.) It also returns a simple bool, treating any errors as false
returns (i.e. "the program shouldn't do anything special.")

Also removed the `controlpersist` check since it is not straightforward
to decide whether this value is set appropriately; infer_create_socket
is a heuristic check, and it's okay for it to be a loose heuristic.