Infer or set whether to create our own socket (#22)
This integrates #9 and #20 into one PR.

We add a `--create-socket` flag, deprecating `--reuse-socket` since the
former seems more intuitive. If `--create-socket` is true, then the
program will create and manage its own temporary SSH control socket; if
it is false, the program will not. If it is unspecified, then the
program will try to parse the output of `ssh -G` to decide whether or
not to create a socket, specifically looking for a `controlmaster auto`
line that would imply that the user is already using their own SSH
connection multiplexing.

This PR also includes a couple other changes:

- We switch to using the `linux-native-sync-persistent` keyring on
  Linux, in the hopes that this will work with what
  `aspect-credential-helper` will be using.
- Increments the version to 0.4.0.

---------

Co-authored-by: Andy Walker <andy@andy.dev>