Remove From<TempDir> from TempSocket
Inlining the code drew my attention to one reason I wrote it the other
way before: from the point we call `dir.into_path()`, we take
responsibility for removing the tempdir ourselves, so we must not
early-return from that point until the TempSocket is created.

Rather than have a separate function, I chose to just call this out with
a comment. Overall I'm happier with this code, there were too many edge
cases for the incoming `TempDir` (chiefly the impossible-to-observe
`keep` flag) for the `From` implementation to have been correct.