Do not staple binary
We would need to make a pkg installer. We don't want to. So we zip the
CLI and notarize the zip (since Apple wants a package format to be
submitted) and then just upload the signed binary itself without the
stapled notarization.

It'll cost a network RTT on first app run. So it goes.