macos codesigning workflow fixes (#38)
This results in a signed, notarized, stapled .pkg file containing a
signed aspect-reauth binary that runs on my laptop. Apparently you can’t
notarize anything other than a .zip, .pkg, .dmg, or .app, and you can’t
just download a signed binary from the internet and run it — it has to
come from a notarized source.

I spent some time trying to go down the road of notarizing the zip that
GitHub creates in its upload-artifact action, but it was shockingly
difficult to get at the actual zip itself rather than the zip’s contents
from within a GitHub action, and besides, zip files cannot have their
notarizations stapled, so you’d need either a network connection or to
know how to run `xattr`.

This also turns on caching for the build workflow since that wound up
being the bottleneck on testing this.

And finally, I decided to move the Format job out of the Build workflow
since formatting does not need to be run as a prerequisite to a release.