build workflow overhaul (#40)
Better caching: one cache only, cache key depends on Rust version. Pins
the rust version used in the build workflow.
[Swatinem/rust-cache](https://github.com/swatinem/rust-cache) needs the
toolchain installation step to be done before the cache is loaded
because it needs the toolchain step to decide on the cache key; by
pinning a rustc version, we do not have that problem.

Builds pull requests with a new `fastbuild` profile, designed to act
like Bazel’s `-c fastbuild`, i.e. no optimizations and minimal debug
info. Pushes to main, and workflow calls for releases, still get built
with `release`.

This also simplifies the build matrix, only specifying the non-redundant
information (so the matrix job names in GitHub will be shorter) and
computing the rest from that.

For the format workflow, I decided not to pin the Rust version; "stable"
seems fine there.