Add a release step to publish to crates.io (#41)
By publishing source-only releases on the [crates.io](https://crates.io)
registry, we allow users to run e.g.:
```sh
env ASPECT_REMOTE=aw-remote-ext.mydomain.example cargo install aspect-reauth@0.7.0
```
and always get a build from the exact same aspect-reauth source code.

This modifies the build step to do a `cargo publish --dry-run` on each
of the supported platforms to make sure that the package can build from
the source we are about to upload. This only happens if the build is
called via `workflow_call`, so in the release workflow. If these dry
runs all succeed, then the release workflow itself then does the actual
publish using the newly added `crates-io` environment.

The `crates-io` environment contains a secret for the cargo registry
token; the documentation I was able to find online suggests that these
tokens should be tied to an individual not an organization, so I’ve gone
ahead and added a token with no expiration to the secrets that has
permissions to publish and yank versions of the [aspect-reauth
crate](https://crates.io/crates/aspect-reauth).

Presently this crate is owned by @bradharris-stairwell and I. If you
should be on that list as well, please contact one of us.