Cache aspect-reauth passwords separately
To reduce the amount of password prompts needed on macos, we cache the
password from Aspect in a separate Keychain entry. Now, if our local
credentials are still valid, no password prompt is required to fetch and
sync them.

If we ever wind up in a situation where the AspectWorkflows credentials
are valid but the aspect-reauth credentials are not, this program will
break until the local credentials are refreshed with `-l` or `-f`.