Reduce number of macos password prompts (#44)
Makes a few modifications, hopefully towards the program not needing
more password re-prompts than necessary.

The two big ones are:
1. Instead of refreshing the credential every time the remote’s is
expired, we check both the local and the remote and only do the work we
need to do, i.e. we only sync the remote credential if it is not
working, and we only re-login if the local credential is not working.
2. We cache the password at a separate Keychain entry owned by this
program. For some reason, otherwise the macos keychain was not playing
nicely between `aspect-reauth` and `aspect-credential-helper`, and we
were still needing to type our password in every time we ran the
program. My best guess is that this is because the credential-helper was
rewriting the credential on every run.

This cache creates the possibility of the two sides of the cache falling
out of sync in strange ways. While this shouldn’t happen as long as you
always use `aspect-reauth` on your laptop (and do not run
`aspect-credential-helper login` directly on your laptop), we try to
defend ourselves against the inevitability that it will happen anyway,
both by falling back to reading the source entry and — as a last-ditch
measure — checking that the remote credential is in fact valid after
we’re done and printing out instructions on what to do if it is not.
(Generally speaking, `--force` should resolve these types of issues.)

To support the increased amount of checks and re-checks without
sacrificing more performance than we need to, this converts the code to
be async using the smol runtime.

This goes ahead and bumps the program to v0.8.0.