Fix ssh `needs_refresh` error log
The log was always printing out `failed to run {helper} on {host}` even
when it was being called locally, which made debugging confusing. Now it
only prints the host if the command came from a `ssh_mux`. The expedient
means of taking an `Option<&SshMux>` was taken to accomplish this.